New uniform_float internal module
